+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
         android:theme="@style/Theme.AppCompat.Light.NoActionBar"
         tools:targetApi="31">
         <activity
-            android:name=".Activity.LandingUser"
+            android:name=".Activity.landingActivity"
             android:exported="false" />
         <activity
             android:name=".Activity.LoginUser"
