+++ app/src/main/java/com/example/text_it/Activity/RegisterUser.kt
@@ -18,7 +18,7 @@ class RegisterUser : AppCompatActivity() {
         backButton.setOnClickListener {
             startActivity(
                 Intent(
-                    this, LandingUser::class.java
+                    this, landingActivity::class.java
                 )
             )
         }
