+++ app/src/main/java/com/example/text_it/Activity/SplashActivity.kt
@@ -3,10 +3,13 @@ package com.example.text_it.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.os.Handler
+import android.view.WindowManager
 import android.widget.ImageView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import com.example.text_it.R
+
+// #1
 
 class SplashActivity : AppCompatActivity() {
     private val SPLASH_TIME_OUT: Long = 3000
@@ -14,6 +17,11 @@ class SplashActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_splash)
+
+        window.setFlags(
+            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS,
+            WindowManager.LayoutParams.FLAG_LAYOUT_NO_LIMITS
+        )
 
         if (checkSelfPermission(android.Manifest.permission.CAMERA) != android.content.pm.PackageManager.PERMISSION_GRANTED) {
             requestPermissions(
@@ -27,14 +35,14 @@ class SplashActivity : AppCompatActivity() {
 
         iv_logo.alpha = 0f
         iv_logo.animate().setDuration(3000).alpha(1f).withEndAction {
-            val i = Intent(this, LandingUser::class.java)
+            val i = Intent(this, landingActivity::class.java)
             startActivity(i)
             overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
             finish()
         }
 
         Handler().postDelayed({
-            val i = Intent(this, LandingUser::class.java)
+            val i = Intent(this, landingActivity::class.java)
             startActivity(i)
             overridePendingTransition(android.R.anim.fade_in, android.R.anim.fade_out)
             finish()
