--- app/src/main/java/com/example/text_it/Fragment/Calls.kt
@@ -1,18 +1,0 @@
-package com.example.text_it.Fragment
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.text_it.R
-
-class Calls : Fragment() {
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_calls, container, false)
-    }
-}
